requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.63
langchain-google-genai==2.0.10
langchain-huggingface==0.2.0
langchain-openai==0.3.18
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
openai==1.82.1
openai-whisper==20240930
minio==7.2.15
python-dotenv==1.1.0
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
pillow==11.2.1
chromadb==1.0.12